+++ ChangeLog.md
@@ -12,10 +12,17 @@ and this project adheres to
 
 ### Changed
 * The `normalize` argument to `freud.density.RDF` is now `normalization_mode`.
-* Rmove zero-padding from arrays in `freud.environment.EnvironmentCluster` and `freud.environment.EnvironmentMotifMatch` and replace with ragged lists of NumPy arrays.
+* The ``neighbors`` argument to ``env_neighbors`` for ``EnvironmentMotifMatch`` class.
+* The ``neighbors`` argument to ``cluster_neighbors`` for ``EnvironmentCluster`` class.
+* Remove zero-padding from arrays in `freud.environment.EnvironmentCluster` and `freud.environment.EnvironmentMotifMatch` and replace with ragged lists of NumPy arrays.
 
 ### Removed
 * The `global_search` flag in `freud.environment.EnvironmentCluster`.
+
+## v2.13.0 -- YYYY-MM-DD
+
+### Added
+* Filter neighborlists with `freud.locality.FilterSANN`.
 
 ## v2.12.1 -- 2022-12-05